Question: Refatorado componente de atividade
- Essa modificação generaliza o componente de atividade, permitindo a
  implementação de atividades do tutorial, que permitem varias respostas
na mesma questão e resultado imediato, ou questões de prova, com logica
diferente
